// hr-backend-api/src/app/employee/dto/employee-dto.ts
export class EmployeeDto {
  id: number;
  name: string;
  surname: string;
  secondName: string;
  sex: boolean;
  address: string;
  phone: string;
  citizenshipId: number;
  birthday: Date;
  marriageId: number;
  militaryId: number;
  statusId: number;
  userId: number;
  employeeAddId: number;
}